refactoring gas cost calculation of forks
