+++ lib/ciri/forks/frontier/cost.rb
@@ -21,7 +21,7 @@ module Ciri
   module Forks
     module Frontier
 
-      module Cost
+      class Cost
 
         include Ciri::EVM::OP
 
@@ -79,115 +79,111 @@ module Ciri
         W_EXTCODE = [EXTCODESIZE]
 
 
-        class << self
-          include Ciri::EVM::OP
-
-          # C(σ,μ,I)
-          # calculate cost of current operation
-          def cost_of_operation(vm)
-            ms = vm.machine_state
-            instruction = vm.instruction
-            w = instruction.get_op(vm.pc)
-            if w == SSTORE
-              cost_of_sstore(vm)
-            elsif w == EXP && ms.get_stack(1, Integer) == 0
-              G_EXP
-            elsif w == EXP && (x = ms.get_stack(1, Integer)) > 0
-              G_EXP + G_EXPBYTE * Utils.ceil_div(x.bit_length, 8)
-            elsif w == CALLDATACOPY || w == CODECOPY || w == RETURNDATACOPY
-              G_VERYLOW + G_COPY * Utils.ceil_div(ms.get_stack(2, Integer), 32)
-            elsif w == EXTCODECOPY
-              G_EXTCODE + G_COPY * Utils.ceil_div(ms.get_stack(3, Integer), 32)
-            elsif (LOG0..LOG4).include? w
-              G_LOG + G_LOGDATA * ms.get_stack(1, Integer) + (w - LOG0) * G_TOPIC
-            elsif w == CALL || w == CALLCODE || w == DELEGATECALL
-              G_CALL
-            elsif w == SELFDESTRUCT
-              cost_of_self_destruct(vm)
-            elsif w == CREATE
-              G_CREATE
-            elsif w == SHA3
-              G_SHA3 + G_SHA3WORD * Utils.ceil_div(ms.get_stack(1, Integer), 32)
-            elsif w == JUMPDEST
-              G_JUMPDEST
-            elsif w == SLOAD
-              G_SLOAD
-            elsif W_ZERO.include? w
-              G_ZERO
-            elsif W_BASE.include? w
-              G_BASE
-            elsif W_VERYLOW.include? w
-              G_VERYLOW
-            elsif W_LOW.include? w
-              G_LOW
-            elsif W_MID.include? w
-              G_MID
-            elsif W_HIGH.include? w
-              G_HIGH
-            elsif W_EXTCODE.include? w
-              G_EXTCODE
-            elsif w == BALANCE
-              G_BALANCE
-            elsif w == BLOCKHASH
-              G_BLOCKHASH
-            else
-              raise "can't compute cost for unknown op #{w}"
-            end
+        # C(σ,μ,I)
+        # calculate cost of current operation
+        def gas_of_operation(vm)
+          ms = vm.machine_state
+          instruction = vm.instruction
+          w = instruction.get_op(vm.pc)
+          if w == SSTORE
+            cost_of_sstore(vm)
+          elsif w == EXP && ms.get_stack(1, Integer) == 0
+            G_EXP
+          elsif w == EXP && (x = ms.get_stack(1, Integer)) > 0
+            G_EXP + G_EXPBYTE * Utils.ceil_div(x.bit_length, 8)
+          elsif w == CALLDATACOPY || w == CODECOPY || w == RETURNDATACOPY
+            G_VERYLOW + G_COPY * Utils.ceil_div(ms.get_stack(2, Integer), 32)
+          elsif w == EXTCODECOPY
+            G_EXTCODE + G_COPY * Utils.ceil_div(ms.get_stack(3, Integer), 32)
+          elsif (LOG0..LOG4).include? w
+            G_LOG + G_LOGDATA * ms.get_stack(1, Integer) + (w - LOG0) * G_TOPIC
+          elsif w == CALL || w == CALLCODE || w == DELEGATECALL
+            G_CALL
+          elsif w == SELFDESTRUCT
+            cost_of_self_destruct(vm)
+          elsif w == CREATE
+            G_CREATE
+          elsif w == SHA3
+            G_SHA3 + G_SHA3WORD * Utils.ceil_div(ms.get_stack(1, Integer), 32)
+          elsif w == JUMPDEST
+            G_JUMPDEST
+          elsif w == SLOAD
+            G_SLOAD
+          elsif W_ZERO.include? w
+            G_ZERO
+          elsif W_BASE.include? w
+            G_BASE
+          elsif W_VERYLOW.include? w
+            G_VERYLOW
+          elsif W_LOW.include? w
+            G_LOW
+          elsif W_MID.include? w
+            G_MID
+          elsif W_HIGH.include? w
+            G_HIGH
+          elsif W_EXTCODE.include? w
+            G_EXTCODE
+          elsif w == BALANCE
+            G_BALANCE
+          elsif w == BLOCKHASH
+            G_BLOCKHASH
+          else
+            raise "can't compute cost for unknown op #{w}"
           end
-
-          def cost_of_memory(i)
-            G_MEMORY * i + (i ** 2) / 512
-          end
-
-          def intrinsic_gas_of_transaction(t)
-            gas = (t.data.each_byte || '').reduce(0) {|sum, i| sum + (i.zero? ? G_TXDATAZERO : G_TXDATANONZERO)}
-            # gas + (t.to.empty? ? G_TXCREATE : 0) + G_TRANSACTION
-            gas + G_TRANSACTION
-          end
-
-          def gas_of_call(vm:, gas:, to:, value:)
-            # TODO handle gas calculation for all categories calls
-            account_exists = vm.account_exist?(to)
-            transfer_gas_fee = value > 0 ? G_CALLVALUE : 0
-            create_gas_fee = !account_exists ? G_NEWACCOUNT : 0
-            extra_gas = transfer_gas_fee + create_gas_fee
-
-            total_fee = gas + extra_gas
-            child_gas_limit = gas + (value > 0 ? G_CALLSTIPEND : 0)
-            [child_gas_limit, total_fee]
-          end
-
-          private
-
-          def cost_of_self_destruct(vm)
-            G_SELFDESTRUCT
-          end
-
-          def cost_of_sstore(vm)
-            ms = vm.machine_state
-            instruction = vm.instruction
-
-            key = ms.get_stack(0, Integer)
-            value = ms.get_stack(1, Integer)
-
-            current_is_empty = vm.fetch(instruction.address, key).zero?
-            value_is_empty = value.nil? || value.zero?
-
-            gas_cost = if current_is_empty && !value_is_empty
-                         G_SSET
-                       else
-                         G_RESET
-                       end
-            gas_refund = if !current_is_empty && value_is_empty
-                           R_SCLEAR
-                         else
-                           0
-                         end
-
-            [gas_cost, gas_refund]
-          end
-
         end
+
+        def gas_of_memory(i)
+          G_MEMORY * i + (i ** 2) / 512
+        end
+
+        def intrinsic_gas_of_transaction(t)
+          gas = (t.data.each_byte || '').reduce(0) {|sum, i| sum + (i.zero? ? G_TXDATAZERO : G_TXDATANONZERO)}
+          # gas + (t.to.empty? ? G_TXCREATE : 0) + G_TRANSACTION
+          gas + G_TRANSACTION
+        end
+
+        def gas_of_call(vm:, gas:, to:, value:)
+          # TODO handle gas calculation for all categories calls
+          account_exists = vm.account_exist?(to)
+          transfer_gas_fee = value > 0 ? G_CALLVALUE : 0
+          create_gas_fee = !account_exists ? G_NEWACCOUNT : 0
+          extra_gas = transfer_gas_fee + create_gas_fee
+
+          total_fee = gas + extra_gas
+          child_gas_limit = gas + (value > 0 ? G_CALLSTIPEND : 0)
+          [child_gas_limit, total_fee]
+        end
+
+        private
+
+        def cost_of_self_destruct(vm)
+          G_SELFDESTRUCT
+        end
+
+        def cost_of_sstore(vm)
+          ms = vm.machine_state
+          instruction = vm.instruction
+
+          key = ms.get_stack(0, Integer)
+          value = ms.get_stack(1, Integer)
+
+          current_is_empty = vm.fetch(instruction.address, key).zero?
+          value_is_empty = value.nil? || value.zero?
+
+          gas_cost = if current_is_empty && !value_is_empty
+                       G_SSET
+                     else
+                       G_RESET
+                     end
+          gas_refund = if !current_is_empty && value_is_empty
+                         R_SCLEAR
+                       else
+                         0
+                       end
+
+          [gas_cost, gas_refund]
+        end
+
       end
 
     end
